Gestió d'estats de solicituds enviades manualment #342
Guardar informació d'enviament a Madrid a una altra taula a BBDD
#343

+++ pinbaladmin-back/src/main/java/org/fundaciobit/pinbaladmin/back/controller/operador/SolicitudLocalDesDeFitxerXmlOperador.java
@@ -228,7 +228,7 @@ public class SolicitudLocalDesDeFitxerXmlOperador extends SolicitudLocalOperador
                 
                 // Gestió del estat Pinbal -- Per defecte no sol·licitat
 				{
-					int estat = Constants.ESTAT_PINBAL_NO_SOLICITAT;
+					Long estat = Constants.ESTAT_PINBAL_NO_SOLICITAT;
 					solicitud.setEstatpinbal(estat);
 				}
